CopWatch/requirements.txt
discord.py
twython